FavoriteToggle component to be unit tested
